+++ src/main/java/roomescape/controller/TimeController.java
@@ -1,0 +1,11 @@
+package roomescape.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/times")
+public class TimeController {
+
+
+}
