--- src/main/java/roomescape/domain/Reservation.java
@@ -1,39 +1,0 @@
-package roomescape.domain;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public class Reservation {
-
-    private final Long id;
-    private final String name;
-    private final LocalDate date;
-    private final LocalTime time;
-
-    public Reservation(String name, LocalDate date, LocalTime time) {
-        this(null, name, date, time);
-    }
-
-    public Reservation(Long id, String name, LocalDate date, LocalTime time) {
-        this.id = id;
-        this.name = name;
-        this.date = date;
-        this.time = time;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public LocalDate getDate() {
-        return date;
-    }
-
-    public LocalTime getTime() {
-        return time;
-    }
-}
