+++ src/main/java/roomescape/domain/reservation/Reservation.java
@@ -1,32 +1,24 @@
-package roomescape.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import roomescape.domain.Reservation;
+package roomescape.domain.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class ReservationResponse {
+public class Reservation {
 
     private final Long id;
-
     private final String name;
-
     private final LocalDate date;
-
-    @JsonFormat(pattern = "HH:mm")
     private final LocalTime time;
 
-    private ReservationResponse(Long id, String name, LocalDate date, LocalTime time) {
+    public Reservation(String name, LocalDate date, LocalTime time) {
+        this(null, name, date, time);
+    }
+
+    public Reservation(Long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
         this.name = name;
         this.date = date;
         this.time = time;
-    }
-
-    public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(reservation.getId(), reservation.getName(),
-                reservation.getDate(), reservation.getTime());
     }
 
     public Long getId() {
