+++ src/main/java/roomescape/domain/reservation/repository/H2ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain.repository;
+package roomescape.domain.reservation.repository;
 
 import java.util.List;
 import javax.sql.DataSource;
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 
 @Repository
 @Primary
