+++ src/main/java/roomescape/domain/reservation/repository/InMemoryReservationRepository.java
@@ -1,10 +1,10 @@
-package roomescape.domain.repository;
+package roomescape.domain.reservation.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 
 @Repository
 public class InMemoryReservationRepository implements ReservationRepository {
