+++ src/main/java/roomescape/dto/reservation/ReservationRequest.java
@@ -1,8 +1,8 @@
-package roomescape.dto;
+package roomescape.dto.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import roomescape.domain.Reservation;
+import roomescape.domain.reservation.Reservation;
 
 public class ReservationRequest {
