+++ src/main/java/roomescape/dto/reservation/ReservationResponse.java
@@ -1,0 +1,46 @@
+package roomescape.dto.reservation;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import roomescape.domain.reservation.Reservation;
+
+public class ReservationResponse {
+
+    private final Long id;
+
+    private final String name;
+
+    private final LocalDate date;
+
+    @JsonFormat(pattern = "HH:mm")
+    private final LocalTime time;
+
+    private ReservationResponse(Long id, String name, LocalDate date, LocalTime time) {
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.time = time;
+    }
+
+    public static ReservationResponse from(Reservation reservation) {
+        return new ReservationResponse(reservation.getId(), reservation.getName(),
+                reservation.getDate(), reservation.getTime());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public LocalTime getTime() {
+        return time;
+    }
+}
